General improvements (#4)
* Add assertions to the tests

This way, we can verify they work

* Require mutable references for tesseract

This way, we prevent changing the tesseract object unless you have a
mutable reference to it

* Remove the libc dependency

It isn't required

* Add myself as an author

* Follow Cargo clippy suggestions

* Investigate memory safety

We should free the image after we pass it into Tesseract.
It makes its own copy.
https://tesseract-ocr.github.io/4.0.0/a01625.html#ga0c4c7f05fd58b3665b123232a05545ad

I'm not sure how to free the memory from get_text. Advice or pull
requests around this will be appreciated.

* Link to tesseract's github

* Bump to 0.2.0

We're allowed to make breaking changes in the before 1.0.0 releases.
This is a breaking change because many of the methods now require a
mutable pointer.