Move unsafe handling code to plumbing module
This has given me the opportunity to review the safety of the unsafe
code - and it was found lacking.

But it's fixed in the plumbing module.

https://github.com/antimatter15/tesseract-rs/issues/16

I've tried to not put naming opinions within the plumbing module.
That is, things within it are named to match the c or c++ libraries of
leptonica and tesseract as much as possible.

This addresses https://github.com/antimatter15/tesseract-rs/issues/17 at
least within the plumbing module.